fix(backup) : 2023. 10 Survivor 2023. 10. 27 백업용
Survival 폴더 참고 !

+++ 0_Scripts/Project/Survivor/01_Attack/2_Missiles/GuidedMissile.cs
@@ -14,7 +14,7 @@ public class GuidedMissile : Missile
     [SerializeField] protected Rigidbody myRigid;
     [SerializeField] protected float moveSpeed;
 
-    [SerializeField] protected short prefabIdx;
+    [SerializeField] protected int prefabIdx;
 
     protected int atk;
 
@@ -39,7 +39,7 @@ public class GuidedMissile : Missile
     }
     */
 
-    public override void Init(Selectable _atker, int _atk, short _prefabIdx)
+    public override void Init(Selectable _atker, int _atk, int _prefabIdx)
     {
 
         atker = _atker.transform;
